move analytics into seperate package
