+++ cmd/earth/main.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"context"
 	"crypto/rand"
-	"crypto/sha256"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
@@ -12,7 +11,6 @@ import (
 	"net/http"
 	_ "net/http/pprof" // enable pprof handlers on net/http listener
 	"os"
-	"os/exec"
 	"os/signal"
 	"os/user"
 	"path/filepath"
@@ -25,6 +23,7 @@ import (
 	"text/tabwriter"
 	"time"
 
+	"github.com/earthly/earthly/analytics"
 	"github.com/earthly/earthly/autocomplete"
 	"github.com/earthly/earthly/buildcontext/provider"
 	"github.com/earthly/earthly/builder"
@@ -39,6 +38,7 @@ import (
 	"github.com/earthly/earthly/earthfile2llb/variables"
 	"github.com/earthly/earthly/llbutil"
 	"github.com/earthly/earthly/secretsclient"
+	"github.com/earthly/earthly/utils"
 
 	"github.com/fatih/color"
 	"github.com/joho/godotenv"
@@ -49,13 +49,11 @@ import (
 	"github.com/moby/buildkit/session/auth/authprovider"
 	"github.com/moby/buildkit/session/sshforward/sshprovider"
 	"github.com/moby/buildkit/util/entitlements"
-	uuid "github.com/nu7hatch/gouuid"
 	"github.com/pkg/errors"
 	"github.com/seehuhn/password"
 	"github.com/urfave/cli/v2"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/sync/errgroup"
-	"gopkg.in/segmentio/analytics-go.v3"
 )
 
 var dotEnvPath = ".env"
@@ -157,7 +155,7 @@ func main() {
 
 	// Load .env into current global env's. This is mainly for applying Earthly settings.
 	// Separate call is made for build args and secrets.
-	if fileExists(dotEnvPath) {
+	if utils.FileExists(dotEnvPath) {
 		err := godotenv.Load(dotEnvPath)
 		if err != nil {
 			fmt.Printf("Error loading dot-env file %s: %s\n", dotEnvPath, err.Error())
@@ -195,7 +193,9 @@ func main() {
 	app.autoComplete()
 
 	exitCode := app.run(ctx, os.Args)
-	app.collectAnalytics(exitCode, time.Since(startTime))
+	if app.cfg == nil || !app.cfg.Global.DisableAnalytics {
+		analytics.CollectAnalytics(Version, GitSha, app.commandName, exitCode, time.Since(startTime))
+	}
 	os.Exit(exitCode)
 }
 
@@ -615,7 +615,7 @@ func (app *earthApp) before(context *cli.Context) error {
 		}
 	}
 
-	if !dirExists(app.cfg.Global.RunPath) {
+	if !utils.DirExists(app.cfg.Global.RunPath) {
 		err := os.MkdirAll(app.cfg.Global.RunPath, 0755)
 		if err != nil {
 			return errors.Wrapf(err, "failed to create run directory %s", app.cfg.Global.RunPath)
@@ -749,12 +749,12 @@ func (app *earthApp) insertBashCompleteEntry() error {
 	}
 	dirPath := filepath.Dir(path)
 
-	if !dirExists(dirPath) {
+	if !utils.DirExists(dirPath) {
 		fmt.Fprintf(os.Stderr, "Warning: unable to enable bash-completion: %s does not exist\n", dirPath)
 		return nil // bash-completion isn't available, silently fail.
 	}
 
-	if fileExists(path) {
+	if utils.FileExists(path) {
 		return nil // file already exists, don't update it.
 	}
 
@@ -815,12 +815,12 @@ func (app *earthApp) insertZSHCompleteEntry() error {
 	path := "/usr/local/share/zsh/site-functions/_earth"
 	dirPath := filepath.Dir(path)
 
-	if !dirExists(dirPath) {
+	if !utils.DirExists(dirPath) {
 		fmt.Fprintf(os.Stderr, "Warning: unable to enable zsh-completion: %s does not exist\n", dirPath)
 		return nil // zsh-completion isn't available, silently fail.
 	}
 
-	if fileExists(path) {
+	if utils.FileExists(path) {
 		return nil // file already exists, don't update it.
 	}
 
@@ -1335,7 +1335,7 @@ func (app *earthApp) actionBuild(c *cli.Context) error {
 	secrets := app.secrets.Value()
 	//interactive debugger settings are passed as secrets to avoid having it affect the cache hash
 	dotEnvMap := make(map[string]string)
-	if fileExists(dotEnvPath) {
+	if utils.FileExists(dotEnvPath) {
 		dotEnvMap, err = godotenv.Read(dotEnvPath)
 		if err != nil {
 			return errors.Wrapf(err, "read %s", dotEnvPath)
@@ -1462,103 +1462,6 @@ func (app *earthApp) newBuildkitdClient(ctx context.Context, opts ...client.Clie
 	return bkClient, nil
 }
 
-func detectCI() string {
-	for k, v := range map[string]string{
-		"GITHUB_WORKFLOW": "github-actions",
-		"CIRCLECI":        "circle-ci",
-		"JENKINS_HOME":    "jenkins",
-		"BUILDKITE":       "buildkite",
-		"DRONE_BRANCH":    "drone",
-		"TRAVIS":          "travis",
-	} {
-		if _, ok := os.LookupEnv(k); ok {
-			return v
-		}
-	}
-
-	// default catch-all
-	if v, ok := os.LookupEnv("CI"); ok {
-		if strings.ToLower(v) == "true" {
-			return "unknown"
-		}
-		return v
-	}
-
-	return "false"
-}
-
-func getRepo() string {
-	cmd := exec.Command("git", "config", "--get", "remote.origin.url")
-	out, err := cmd.Output()
-	if err == nil {
-		return string(out)
-	}
-
-	for _, k := range []string{
-		"GITHUB_REPOSITORY",
-		"CIRCLE_PROJECT_REPONAME",
-		"GIT_URL",
-		"BUILDKITE_REPO",
-		"DRONE_REPO",
-		"TRAVIS_REPO_SLUG",
-	} {
-		if v, ok := os.LookupEnv(k); ok {
-			return v
-		}
-	}
-
-	for _, e := range os.Environ() {
-		pair := strings.SplitN(e, "=", 2)
-		if len(pair) == 2 {
-			if strings.Contains(pair[1], "git") {
-				return pair[1]
-			}
-		}
-	}
-
-	return "unknown"
-}
-
-func getRepoHash() string {
-	repo := getRepo()
-	if repo == "unknown" || repo == "" {
-		return repo
-	}
-	return fmt.Sprintf("%x", sha256.Sum256([]byte(repo)))
-}
-
-func (app *earthApp) collectAnalytics(exitCode int, realtime time.Duration) {
-	if app.cfg != nil && app.cfg.Global.DisableAnalytics {
-		return
-	}
-	installID, err := getInstallID()
-	if err != nil {
-		installID = "unknown"
-	}
-	segmentClient := analytics.New("RtwJaMBswcW3CNMZ7Ops79dV6lEZqsXf")
-	segmentClient.Enqueue(analytics.Track{
-		Event:  "cli-" + app.commandName,
-		UserId: installID,
-		Properties: analytics.NewProperties().
-			Set("version", Version).
-			Set("gitsha", GitSha).
-			Set("exitcode", exitCode).
-			Set("ci", detectCI()).
-			Set("repo", getRepoHash()).
-			Set("realtime", realtime.Seconds()),
-	})
-	done := make(chan bool, 1)
-	go func() {
-		segmentClient.Close()
-		done <- true
-	}()
-	select {
-	case <-time.After(time.Millisecond * 500):
-	case <-done:
-	}
-
-}
-
 func processSecrets(secrets []string, dotEnvMap map[string]string) (map[string][]byte, error) {
 	finalSecrets := make(map[string][]byte)
 	for k, v := range dotEnvMap {
@@ -1589,59 +1492,8 @@ func defaultConfigPath() string {
 
 	oldConfig := filepath.Join(homeDir, ".earthly", "config.yaml")
 	newConfig := filepath.Join(homeDir, ".earthly", "config.yml")
-	if fileExists(oldConfig) && !fileExists(newConfig) {
+	if utils.FileExists(oldConfig) && !utils.FileExists(newConfig) {
 		return oldConfig
 	}
 	return newConfig
-}
-
-func getInstallID() (string, error) {
-	installID, ok := os.LookupEnv("EARTHLY_INSTALL_ID")
-	if ok {
-		return installID, nil
-	}
-
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get user home dir")
-	}
-
-	path := filepath.Join(homeDir, ".earthly", "install_id")
-	if !fileExists(path) {
-
-		u, err := uuid.NewV4()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to generate uuid")
-		}
-
-		ID := u.String()
-
-		err = ioutil.WriteFile(path, []byte(ID), 0644)
-		if err != nil {
-			return "", errors.Wrapf(err, "failed to write %q", path)
-		}
-		return ID, nil
-	}
-
-	s, err := ioutil.ReadFile(path)
-	if err != nil {
-		return "", errors.Wrapf(err, "failed to read %q", path)
-	}
-	return string(s), nil
-}
-
-func fileExists(filename string) bool {
-	info, err := os.Stat(filename)
-	if os.IsNotExist(err) {
-		return false
-	}
-	return !info.IsDir()
-}
-
-func dirExists(filename string) bool {
-	info, err := os.Stat(filename)
-	if os.IsNotExist(err) {
-		return false
-	}
-	return info.IsDir()
 }
