+++ utils/utils.go
@@ -1,0 +1,23 @@
+package utils
+
+import (
+	"os"
+)
+
+// FileExists returns true if the file exists
+func FileExists(filename string) bool {
+	info, err := os.Stat(filename)
+	if os.IsNotExist(err) {
+		return false
+	}
+	return !info.IsDir()
+}
+
+// DirExists returns true if the directory exists
+func DirExists(filename string) bool {
+	info, err := os.Stat(filename)
+	if os.IsNotExist(err) {
+		return false
+	}
+	return info.IsDir()
+}
